[ALPS02312673] mtee: fix driver build fail
[Detail] Variable casting and CONFIG_COMPAT feature cause failure.

[Solution] Fix the problem make building fail.

[Feature] MTK In-House TEE (Trusted Execution Environment)

MTK-Commit-Id: 46f6e088c8836e4e44eb1fab7da85d59b032f6a1

CR-Id: ALPS02312673

+++ drivers/misc/mediatek/trustzone/kree_mem.c
@@ -97,7 +97,7 @@ TZ_RESULT kree_register_sharedmem(KREE_SESSION_HANDLE session,
 	TZ_RESULT ret;
 
 	p[0].value.a = (unsigned long)start;
-	p[0].value.b = (unsigned long)start >> 32;
+	p[0].value.b = (unsigned long long)(unsigned long)start >> 32;
 	p[1].value.a = size;
 	p[2].mem.buffer = map_p;
 	if (map_p != NULL)
